requirements.txt
ollama
langchain==0.2.16
langchain-community==0.2.16
langgraph==0.2.28
lancedb==0.13.0
fastapi==0.115.0
uvicorn==0.31.0
slack-bolt==1.21.0
python-dotenv==1.0.1
langfuse==2.53.6
pydantic==2.9.2
pydantic-settings==2.5.2
structlog==24.4.0
pandas>=2.0.0
pyarrow==17.0.0
httpx==0.27.2
pytest==8.3.3
mcp==1.1.2
